refactor(core): move the read core into module functions under core/
