+++ packages/envapt/src/core/EnvironmentMethods.ts
@@ -1,99 +1,19 @@
+import { determineEnvironment } from './engine';
 import { EnvapterBase } from './EnvapterBase';
+import { Environment } from './Environment';
 import { state } from './state';
-import { debugWarn } from '../infra/Debug';
-import { EnvaptError, EnvaptErrorCodes } from '../infra/Error';
-
-import type { EnvProfile } from '../types';
-
-/**
- * Environment types supported by Envapter
- *
- * The following keys are checked in order until the first with a non-empty value is found, or defaulting to development if none are set:
- * - `ENVIRONMENT`
- * - `ENV`
- * - `NODE_ENV`
- * - `MODE`
- *
- * @public
- */
-export enum Environment {
-    /** The default when no environment variable names a known environment. */
-    Development,
-    /** Selected when an environment variable reads `staging`. */
-    Staging,
-    /** Selected when an environment variable reads `production`. */
-    Production,
-    /** Selected when an environment variable reads `test`. */
-    Test
-}
-
-// Keys carrying the environment name, highest precedence first. Checked in order until the first with a non-empty value is found, or defaulting to development if none are set.
-const ENV_KEYS = ['ENVIRONMENT', 'ENV', 'NODE_ENV', 'MODE'] as const;
-
-function parseEnvironment(raw: string): Environment | undefined {
-    switch (raw.toLowerCase()) {
-        case 'production':
-            return Environment.Production;
-        case 'staging':
-            return Environment.Staging;
-        case 'test':
-            return Environment.Test;
-        case 'development':
-            return Environment.Development;
-        default:
-            return undefined;
-    }
-}
 
 /**
  * Mixin for environment detection and checking methods
  * @internal
  */
 export class EnvironmentMethods extends EnvapterBase {
-    protected static determineEnvironment(env?: string | Environment): void {
-        if (typeof env === 'number') {
-            state.environment = env;
-            state.environmentExplicitlySet = true;
-            return;
-        }
-        if (typeof env === 'string') {
-            state.environment = parseEnvironment(env) ?? Environment.Development;
-            state.environmentExplicitlySet = true;
-            return;
-        }
-
-        const raw = this.firstEnvKeyValue((key) => {
-            const value = this.config.get(key);
-            return typeof value === 'string' ? value : undefined;
-        });
-        if (raw === undefined) {
-            debugWarn(`no environment set (looked for ${ENV_KEYS.join(', ')}); defaulting to development`);
-            state.environment = Environment.Development;
-            return;
-        }
-        const parsed = parseEnvironment(raw);
-        if (parsed === undefined) {
-            debugWarn(`unrecognized environment "${raw}"; defaulting to development`);
-            state.environment = Environment.Development;
-            return;
-        }
-        state.environment = parsed;
-    }
-
-    private static firstEnvKeyValue(read: (key: string) => string | undefined): string | undefined {
-        for (const key of ENV_KEYS) {
-            const value = read(key);
-            if (value !== undefined && value.length > 0) return value;
-        }
-        return undefined;
-    }
-
     /**
      * Get the current application environment
      */
     static get environment(): Environment {
         if (state.environment === undefined) {
-            this.determineEnvironment();
+            determineEnvironment();
         }
         return state.environment as Environment;
     }
@@ -102,7 +22,7 @@ export class EnvironmentMethods extends EnvapterBase {
      * Set the application environment. Accepts either Environment enum or string value.
      */
     static set environment(env: string | Environment) {
-        this.determineEnvironment(env);
+        determineEnvironment(env);
     }
 
     /**
@@ -116,7 +36,7 @@ export class EnvironmentMethods extends EnvapterBase {
      * @see {@link EnvironmentMethods.environment}
      */
     set environment(env: string | Environment) {
-        EnvironmentMethods.determineEnvironment(env);
+        determineEnvironment(env);
     }
 
     /**
@@ -173,90 +93,5 @@ export class EnvironmentMethods extends EnvapterBase {
      */
     get isTest(): boolean {
         return EnvironmentMethods.environment === Environment.Test;
-    }
-
-    protected static override refreshCache(): void {
-        // If the env was inferred (not user-set), reset it so re-hydration re-determines
-        // from current state. If the user explicitly set Envapter.environment = X, preserve
-        // that value through the refresh. The immediate re-hydration inside super.refreshCache()
-        // will use it for cascade selection.
-        if (!state.environmentExplicitlySet) state.environment = undefined;
-        super.refreshCache();
-    }
-
-    /**
-     * Reads the source's raw vars (not `this.config`, which would recurse: cascade selection runs
-     * before the `.env` load). `Envapter.environment` may differ post-load if a file sets `ENVIRONMENT`.
-     * @internal
-     */
-    protected static getCascadeEnvironment(): Environment {
-        if (state.environment !== undefined) return state.environment;
-
-        const vars = state.source.readVars();
-        const raw = this.firstEnvKeyValue((key) => vars[key]);
-        return raw === undefined ? Environment.Development : (parseEnvironment(raw) ?? Environment.Development);
-    }
-
-    /**
-     * Build the dotenv-flow cascade for a given environment, in dotenv first-wins precedence
-     * order (highest precedence first). Missing files are silently filtered.
-     *
-     * Precedence is **most-specific-wins** (matches Vite / Astro / Vocs):
-     *   `.env.${env}.local` \> `.env.${env}` \> `.env.local` \> `.env`
-     *
-     * This differs from dotenv-flow / Next.js convention which puts `.env.local` above
-     * `.env.${env}`. We chose the most-specific-wins order so committed env-specific files
-     * (`.env.production`) are authoritative for that environment regardless of whether a
-     * stray `.env.local` is present.
-     * @internal
-     */
-    protected static buildCascadePaths(env: Environment): string[] {
-        const envName = Environment[env].toLowerCase();
-        return [`.env.${envName}.local`, `.env.${envName}`, '.env.local', '.env']
-            .map((name) => this.resolveAgainstBase(name))
-            .filter((p) => this.sourceFileExists(p));
-    }
-
-    private static normalizeProfilePaths(profile: EnvProfile | undefined): string[] {
-        if (!profile) return [];
-        return Array.isArray(profile.paths) ? profile.paths : [profile.paths];
-    }
-
-    /**
-     * Override the base implementation to layer the dotenv-flow cascade + any
-     * `Envapter.configureProfiles` overrides on top of `state.envPaths` when the user has NOT
-     * explicitly set `envPaths`.
-     *
-     * Precedence (passed to dotenv with first-wins semantics):
-     *   1. profile-configured paths for the current env (if any)
-     *   2. `.env.${env}.local`
-     *   3. `.env.${env}`
-     *   4. `.env.local`
-     *   5. `.env`
-     *
-     * If `useDefaults: false` is set on the profiles config, only (1) is loaded, no cascade.
-     * If `envPaths` was explicitly set, only `envPaths` is loaded (everything else ignored).
-     * @internal
-     */
-    protected static override resolveEffectivePaths(): string[] {
-        if (state.envPathsExplicitlySet) return state.envPaths.map((p) => this.resolveAgainstBase(p));
-
-        const env = this.getCascadeEnvironment();
-        const profileEntry = state.profiles?.[env];
-        const profilePaths = this.normalizeProfilePaths(profileEntry);
-
-        // Validate that explicitly configured profile paths exist for the active env.
-        if (profilePaths.length > 0) {
-            const missing = profilePaths.filter((p) => !this.sourceFileExists(this.resolveAgainstBase(p)));
-            if (missing.length > 0) {
-                throw new EnvaptError(
-                    EnvaptErrorCodes.EnvFilesNotFound,
-                    `Environment file not found at path: ${missing.join(', ')}`
-                );
-            }
-        }
-
-        const cascade = state.profiles?.useDefaults === false ? [] : this.buildCascadePaths(env);
-        return [...profilePaths.map((p) => this.resolveAgainstBase(p)), ...cascade];
     }
 }
