+++ packages/envapt/src/core/PrimitiveMethods.ts
@@ -1,62 +1,12 @@
-import { BuiltInConverters, ValueConverter } from '../converters';
+import { Primitive, readPrimitive } from './engine';
 import { EnvironmentMethods } from './EnvironmentMethods';
-import { TemplateResolver } from '../engine/TemplateResolver';
-import { debugWarn } from '../infra/Debug';
 
 import type { ConditionalReturn, EnvKeyInput } from '../types';
-import type { EnvapterService } from '../types/Env';
 
 /**
  * @internal
  */
-enum Primitive {
-    String,
-    Number,
-    Boolean,
-    BigInt,
-    Symbol
-}
-
-/**
- * @internal
- */
-export class PrimitiveMethods extends EnvironmentMethods implements EnvapterService {
-    private static readonly service = new PrimitiveMethods();
-    protected static readonly templateResolver: TemplateResolver = new TemplateResolver(PrimitiveMethods.service);
-    protected static readonly valueConverter: ValueConverter = new ValueConverter(PrimitiveMethods.service);
-
-    protected static override resolveForMirror(key: string, value: string): string {
-        return this.templateResolver.resolveTemplate(key, value);
-    }
-
-    private static _get<EnvVarReturnType, DefaultType extends EnvVarReturnType | undefined = undefined>(
-        key: EnvKeyInput,
-        type: Primitive,
-        def?: DefaultType
-    ): ConditionalReturn<EnvVarReturnType, DefaultType> {
-        const { key: resolvedKey, value } = this.resolveKeyInput(key);
-        if (this.treatAsMissing(value)) {
-            if (def !== undefined) debugWarn(`${resolvedKey} is missing or empty, using fallback ${String(def)}`);
-            else debugWarn(`${resolvedKey} is missing or empty`);
-            return def as ConditionalReturn<EnvVarReturnType, DefaultType>;
-        }
-        const rawVal = value as string | number | boolean | undefined;
-
-        const parsed = this.templateResolver.resolveTemplate(resolvedKey, String(rawVal));
-
-        let result: EnvVarReturnType;
-        if (type === Primitive.Number) result = BuiltInConverters.number(parsed, def as number) as EnvVarReturnType;
-        else if (type === Primitive.Boolean)
-            result = BuiltInConverters.boolean(parsed, def as boolean) as EnvVarReturnType;
-        else if (type === Primitive.BigInt)
-            result = BuiltInConverters.bigint(parsed, def as bigint) as EnvVarReturnType;
-        else if (type === Primitive.Symbol)
-            result = BuiltInConverters.symbol(parsed, def as symbol) as EnvVarReturnType;
-        else result = BuiltInConverters.string(parsed, def as string) as EnvVarReturnType;
-
-        return result;
-    }
-
+export class PrimitiveMethods extends EnvironmentMethods {
     /**
      * Get a string environment variable with optional fallback.
      * Supports template variable resolution using `${VAR}` syntax.
@@ -66,7 +16,7 @@ export class PrimitiveMethods extends EnvironmentMethods implements EnvapterServ
         key: EnvKeyInput,
         def?: Default
     ): ConditionalReturn<string, Default> {
-        return this._get(key, Primitive.String, def);
+        return readPrimitive(key, Primitive.String, def);
     }
 
     /**
@@ -76,7 +26,7 @@ export class PrimitiveMethods extends EnvironmentMethods implements EnvapterServ
         key: EnvKeyInput,
         def?: Default
     ): ConditionalReturn<string, Default> {
-        return PrimitiveMethods._get(key, Primitive.String, def);
+        return readPrimitive(key, Primitive.String, def);
     }
 
     /**
@@ -88,7 +38,7 @@ export class PrimitiveMethods extends EnvironmentMethods implements EnvapterServ
         key: EnvKeyInput,
         def?: Default
     ): ConditionalReturn<number, Default> {
-        return this._get(key, Primitive.Number, def);
+        return readPrimitive(key, Primitive.Number, def);
     }
 
     /**
@@ -98,7 +48,7 @@ export class PrimitiveMethods extends EnvironmentMethods implements EnvapterServ
         key: EnvKeyInput,
         def?: Default
     ): ConditionalReturn<number, Default> {
-        return PrimitiveMethods._get(key, Primitive.Number, def);
+        return readPrimitive(key, Primitive.Number, def);
     }
 
     /**
@@ -110,7 +60,7 @@ export class PrimitiveMethods extends EnvironmentMethods implements EnvapterServ
         key: EnvKeyInput,
         def?: Default
     ): ConditionalReturn<boolean, Default> {
-        return this._get(key, Primitive.Boolean, def);
+        return readPrimitive(key, Primitive.Boolean, def);
     }
 
     /**
@@ -120,7 +70,7 @@ export class PrimitiveMethods extends EnvironmentMethods implements EnvapterServ
         key: EnvKeyInput,
         def?: Default
     ): ConditionalReturn<boolean, Default> {
-        return PrimitiveMethods._get(key, Primitive.Boolean, def);
+        return readPrimitive(key, Primitive.Boolean, def);
     }
 
     /**
@@ -132,7 +82,7 @@ export class PrimitiveMethods extends EnvironmentMethods implements EnvapterServ
         key: EnvKeyInput,
         def?: Default
     ): ConditionalReturn<bigint, Default> {
-        return this._get(key, Primitive.BigInt, def);
+        return readPrimitive(key, Primitive.BigInt, def);
     }
 
     /**
@@ -142,7 +92,7 @@ export class PrimitiveMethods extends EnvironmentMethods implements EnvapterServ
         key: EnvKeyInput,
         def?: Default
     ): ConditionalReturn<bigint, Default> {
-        return PrimitiveMethods._get(key, Primitive.BigInt, def);
+        return readPrimitive(key, Primitive.BigInt, def);
     }
 
     /**
@@ -154,7 +104,7 @@ export class PrimitiveMethods extends EnvironmentMethods implements EnvapterServ
         key: EnvKeyInput,
         def?: Default
     ): ConditionalReturn<symbol, Default> {
-        return this._get(key, Primitive.Symbol, def);
+        return readPrimitive(key, Primitive.Symbol, def);
     }
 
     /**
@@ -164,6 +114,6 @@ export class PrimitiveMethods extends EnvironmentMethods implements EnvapterServ
         key: EnvKeyInput,
         def?: Default
     ): ConditionalReturn<symbol, Default> {
-        return PrimitiveMethods._get(key, Primitive.Symbol, def);
+        return readPrimitive(key, Primitive.Symbol, def);
     }
 }
