+++ packages/envapt/src/core/index.ts
@@ -1,2 +1,2 @@
-export { Environment } from './EnvironmentMethods';
+export { Environment } from './Environment';
 export { AdvancedMethods } from './AdvancedMethods';
